requirements.txt
# Core LangChain - Updated for Python 3.13 compatibility
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchainhub>=0.1.21
langgraph>=0.2.0
langsmith>=0.2.0

# LLM Providers
openai>=1.50.0  # For OpenRouter API compatibility

# Vector Stores & Embeddings
chromadb>=0.5.0
langchain-chroma>=1.0.0
sentence-transformers>=3.0.0

# Document Processing
pypdf>=5.0.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.6

# Web Tools
requests>=2.32.0
duckduckgo-search>=6.0.0
ddgs>=1.1.0

# API Framework (for later)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Utilities
python-dotenv>=1.0.0

# Jupyter Support
ipykernel>=6.29.0
jupyter>=1.1.0
